+++ MicrofluidSimulator/MicrofluidSimulator/SimulatorCode/DataTypes/TemperatureSensor.cs
@@ -4,7 +4,7 @@
     {
 
         //int valueTemperature;
-        public TemperatureSensor(string name, int ID, int sensorID, string type, int positionX, int positionY, int sizeX, int sizeY, int valueTemperature, int electrodeId) 
+        public TemperatureSensor(string name, int ID, int sensorID, string type, int positionX, int positionY, int sizeX, int sizeY, float valueTemperature, int electrodeId) 
             : base(name, ID, sensorID, type, positionX, positionY, sizeX, sizeY, electrodeId)
         {
             this.valueTemperature = valueTemperature;
